refactor: 테이블 컬럼 타입 변경 및 BigDecimal 로직 제거

+++ src/main/java/kitchenpos/menu/domain/MenuV2.java
@@ -56,7 +56,7 @@ public class MenuV2 {
     }
 
     public MenuResponse toMenuResponse() {
-        return new MenuResponse(this.id, this.name, BigDecimal.valueOf(this.price), this.menuGroupId, this.menuProducts);
+        return new MenuResponse(this.id, this.name, this.price, this.menuGroupId, this.menuProducts);
     }
 
     @Override
