+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,19 +1,19 @@
 package kitchenpos.menu.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.domain.MenuV2;
 
 public class MenuRequest {
     private String name;
-    private BigDecimal price;
+    private Long price;
     private Long menuGroupId;
     private List<MenuProductRequest> menuProducts;
 
     public MenuRequest() {
     }
 
-    public MenuRequest(String name, BigDecimal price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
+    public MenuRequest(String name, Long price, Long menuGroupId,
+                       List<MenuProductRequest> menuProducts) {
         this.name = name;
         this.price = price;
         this.menuGroupId = menuGroupId;
@@ -24,7 +24,7 @@ public class MenuRequest {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 
@@ -37,6 +37,6 @@ public class MenuRequest {
     }
 
     public MenuV2 toMenu() {
-        return new MenuV2(null, this.name, this.price.longValue(), this.menuGroupId, null);
+        return new MenuV2(null, this.name, this.price, this.menuGroupId, null);
     }
 }
